Fixed more bugs during testing. Still have some issues with more bugs.
