+++ indeca/pipeline/iteration.py
